电检报告完善

+++ ruoyi-admin/src/main/java/com/ruoyi/electrical/report/dto/OwnerUnitInfo.java
@@ -82,5 +82,5 @@ public class OwnerUnitInfo {
 	/**
 	 * A级隐患数
 	 */
-	private Integer dangers;
+	private Long dangers;
 }
